* Faction rating and combat rating are now continuous.

+++ src/faction.h
@@ -21,10 +21,10 @@ char* faction_longname( int f );
 glTexture* faction_logoSmall( int f );
 
 /* player stuff */
-void faction_modPlayer( int f, int mod );
-void faction_modPlayerRaw( int f, int mod );
-int faction_getPlayer( int f );
-char* faction_getStanding( int mod );
+void faction_modPlayer( int f, double mod );
+void faction_modPlayerRaw( int f, double mod );
+double faction_getPlayer( int f );
+char* faction_getStanding( double mod );
 glColour* faction_getColour( int f );
 
 /* works with only factions */
